made gui_nochat_object object oriented
